+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Next Release
 
  - Print more info when original test fails #549
+ - Reporting JSON to Mutation Testing Elements (initial support) #517, #507
 
 ## [0.4.0] - 11 Aug 2019